+++ Demos/GenericMessage/GenericMessageDemo.cpp
@@ -77,15 +77,6 @@ int main(int argc, char** argv)
     std::cout << "Creating GenericAdapter for participant=" << participantName << " in domain " << domainId << std::endl;
     auto comAdapter = ib::CreateFastRtpsComAdapter(ibConfig, participantName, domainId);
 
-    // Create a SyncMaster if the participant is configured to do so.
-    // NB: New SyncMaster Interface is subject to changes (cf. AFTMAGT-124)
-    auto& participantConfig = ib::cfg::get_by_name(ibConfig.simulationSetup.participants, participantName);
-    if (participantConfig.isSyncMaster)
-    {
-        comAdapter->CreateSyncMaster();
-        std::cout << "Created SyncMaster at Participant: " << participantName << std::endl;
-    }
-
     // Set an Init Handler
     auto&& participantController = comAdapter->GetParticipantController();
     participantController->SetInitHandler([&participantName](auto initCmd) {