+++ IntegrationBus/CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Added
 - Strict sync, i.e., calling wait_for_all_acked() between ticks, is now
   configurable via the IbConfig.json: SimulationSetup/TimeSync/SyncPolicy.
+- If a participant is configured as SyncMaster, the corresponding
+  ComAdapter will now automatically create the SyncMaster instance.
 ### Removed
 - SimulationSetup/TimeSync/SyncType has been removed from the IbConfig.json as
   the SyncType can now be configured per participant.
@@ -17,6 +19,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
   Old: ib::cfg::SimulationSetupBuilder::SetSyncType(SyncType).
 - ComAdapter is now configured automatically according to SyncPolicy. Only in
   strict mode, wait_for_all_acked() is used and a short heartbeat period is used.
+- IComAdapter::CreateSyncMaster() was renamed to IComAdapter::GetSyncMaster() since the
+  SyncMaster is automatically instantiated by the FastRtpsComAdapter if configured.
 
 ### Fixed
 - SystemMonitor was made more robust to race conditions that could lead to a IB